Simplify manual test script
